pyproject.toml
[project]
name = "rag-api"
version = "0.1.0"
description = "RAG API with document ingestion and conversational chat"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "sqlmodel>=0.0.32",
    "qdrant-client>=1.0.0",
    "optimum[onnxruntime]>=1.9.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "onnxruntime>=1.15.0",
    "pdfplumber",
    "redis>=4.0.0",
    "spacy>=3.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.2.1",
    "pip>=26.0.1",
    "chardet>=5.2.0",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.31.0",
    "sqlalchemy[asyncio]>=2.0.46",
]
